Original and Edited Code
This commit contains the original code we used for the data analysis. The code was cleaned out when creating a shiny app. The modified code contains only codes for the graphs. To run the file titled "Final Code Edited for Shiny.R" you will need to download the RDA file as it contains the dataframes for the graphs.